Fixed Firebase functions and AuthProvider context
