+++ .gitignore
@@ -22,5 +22,4 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-
-/package-lock.json
+.env
