+++ client/.env
@@ -1,3 +1,9 @@
+REACT_APP_FIREBASE_KEY = "AIzaSyDZiej5JvC0YD9QCkYaU9Vlg5IzSNkoh8Y"
+REACT_APP_FIREBASE_DOMAIN = "music-streaming-service-58b66.firebaseapp.com"
+REACT_APP_FIREBASE_PROJECT_ID = "music-streaming-service-58b66"
+REACT_APP_FIREBASE_STORAGE_BUCKET = "music-streaming-service-58b66.appspot.com"
+REACT_APP_FIREBASE_SENDER_ID = "439038259088"
+REACT_APP_FIREBASE_APP_ID = "1:439038259088:web:5d049c65fce7b0b4574d15"
 REACT_APP_CLIENT_ID = "1344e7c26ce641739773990965926b6a"
 REACT_APP_CLIENT_SECRET = "29f0184193bb435a84925e757b221731"
 REACT_APP_NEW_RELEASE = "https://api.spotify.com/v1/browse/new-releases"
