+++ client/package.json
@@ -14,7 +14,7 @@
     "@mui/system": "^5.10.16",
     "axios": "^1.2.1",
     "dotenv": "^16.0.3",
-    "firebase": "^9.14.0",
+    "firebase": "^9.15.0",
     "firebaseui": "^6.0.2",
     "formik": "^2.2.9",
     "fs": "^0.0.1-security",
