+++ client/src/components/PrivateRoute.js
@@ -1,0 +1,10 @@
+import { Navigate, Outlet } from "react-router-dom";
+import React, { useContext } from "react";
+import { AuthContext } from "../firebase/Auth";
+
+const PrivateRoute = () => {
+  const { currentUser } = useContext(AuthContext);
+  return currentUser ? <Outlet /> : <Navigate to="/signin" />;
+};
+
+export default PrivateRoute;
