+++ client/src/firebase/Auth.js
@@ -1,31 +1,21 @@
-// import React, { useState, useEffect } from "react";
-// import Fire
-// export const AuthContext = React.createContext();
+import React, { useState, useEffect } from "react";
+import firebaseApp from "./Firebase";
+import { getAuth, onAuthStateChanged } from "firebase/auth";
+export const AuthContext = React.createContext();
 
-// export const AuthProvider = ({ children }) => {
-//   const [currentUser, setCurrentUser] = useState(null);
-//   const [loadingUser, setLoadingUser] = useState(true);
+export const AuthProvider = ({ children }) => {
+  const [currentUser, setCurrentUser] = useState(null);
 
-//   useEffect(() => {
-//     firebaseApp.auth().onAuthStateChanged((user) => {
-//       setCurrentUser(user);
-//       setLoadingUser(false);
-//     });
-//   }, []);
+  useEffect(() => {
+    const auth = getAuth(firebaseApp);
+    onAuthStateChanged(auth, (user) => {
+      setCurrentUser(user);
+    });
+  }, []);
 
-//   if (loadingUser) {
-//     return (
-//       <div>
-//         <h1>Loading....Loading....Loading....Loading....Loading....</h1>
-//       </div>
-//     );
-//   }
-
-//   return (
-//     <AuthContext.Provider value={{ currentUser }}>
-//       {children}
-//     </AuthContext.Provider>
-//   );
-// };
-
-// export default Auth;
+  return (
+    <AuthContext.Provider value={{ currentUser }}>
+      {children}
+    </AuthContext.Provider>
+  );
+};
