+++ client/src/pages/NewReleases.js
@@ -1,18 +1,16 @@
 import React, { useEffect, useState } from "react";
 import axios from "axios";
-import { Link,useParams } from "react-router-dom";
+
 import {
   Card,
   CardActions,
-  CardContent,
   CardHeader,
   CardMedia,
   Grid,
   makeStyles,
   Button,
 } from "@material-ui/core";
-// import { Default } from "react-toastify/dist/utils";
-import DefaultImage from "../img/DefaultImage.jpeg";
+import { Link } from "react-router-dom";
 
 const useStyles = makeStyles({
   card: {
@@ -45,29 +43,30 @@ const useStyles = makeStyles({
   },
 });
 
-const AlbumSong = () => {
-
+const NewReleases = () => {
+  const [musicAlbums, setMusicAlbums] = useState([]);
   const classes = useStyles();
-  //to get data of particular albums
-  const {AlbumId}=useParams();
-
-
-  const [trackAlbums, setTrackAlbums] = useState();
   const [loading, setLoading] = useState(true);
   const [found, setFound] = useState(false);
+  const [playlistData, setPlayListData] = useState({});
 
-  const addToPlaylist = async (trackId) => {
+  console.log(
+    "accessstoken from new release=",
+    window.localStorage.getItem("token"),
+  );
+
+  const addToPlaylist = async (albumId) => {
     try {
       const { data } = await axios.get(
-        `http://localhost:3008/playlist/6393c998ba7131648ed117dc/${trackId}`
+        `http://localhost:3008/playlist/6393c998ba7131648ed117dc/${albumId}`,
       );
+      setPlayListData(data);
     } catch (error) {
       console.log("error", error);
     }
   };
 
-  const getAlbums = async () => {
-
+  const getNewMusicAlbumReleases = async () => {
     const requestInit = {
       headers: {
         Authorization: `Bearer ${window.localStorage.getItem("token")}`,
@@ -77,18 +76,16 @@ const AlbumSong = () => {
 
     try {
       const response = await axios.get(
-        `${process.env.REACT_APP_URL_ALBUMS_ID_TRACKS}/${AlbumId}`,
+        `${process.env.REACT_APP_NEW_RELEASE}`,
         requestInit,
       );
-
       console.log("we get response");
-      console.log("data=",response.data);
-
+      console.log(response);
       setLoading(false);
       setFound(true);
-      setTrackAlbums(response.data);
-
-   
+      setMusicAlbums(response.data.albums.items);
+      console.log("THIS IS URL DATA", response);
+      // console.log(musicAlbums);
     } catch (error) {
       setFound(false);
       setLoading(false);
@@ -97,18 +94,39 @@ const AlbumSong = () => {
   };
 
   useEffect(() => {
-    getAlbums();
-  }, [AlbumId]);
+    getNewMusicAlbumReleases();
+  }, []);
 
-  const buildCard = (track) => {
+  const buildCard = (album) => {
     return (
-      <Grid item xs={12} sm={6} md={4} lg={3} xl={2} key={track?.id}>
+      <Grid item xs={12} sm={6} md={4} lg={3} xl={2} key={album?.id}>
         <Card className={classes.card} variant="outlined">
-          
-             <CardHeader className={classes.titleHead} title={track?.id} />
-              <CardHeader className={classes.titleHead} title={track?.name} />
-        <br/>
-          <Button  onClick={() => addToPlaylist(track?.id)}>Add To PlayList</Button>
+          <CardActions>
+            <Link to={`/${album?.id}`}>
+              <CardHeader className={classes.titleHead} title={album?.name} />
+
+              <CardMedia
+                className={classes.media}
+                component="img"
+                image={album?.images[0]?.url}
+                title="character image"
+              />
+            </Link>
+          </CardActions>
+          {/* <Button
+            className={classes.button}
+            onClick={() => addToPlaylist(album.id)}
+          >
+            Add
+          </Button> */}
+          <br />
+          <Button>
+            {album?.tracks?.itmes[0]?.external_urls.spotify}
+            {/* {album?.disc_number} */}
+            {/* {album?.artists[0]?.map((i) => {
+              return i.external_urls;
+            })} */}
+          </Button>
         </Card>
       </Grid>
     );
@@ -126,37 +144,18 @@ const AlbumSong = () => {
   } else
     return (
       <div>
-        <h1>{" tracks"}</h1>
-       
-  
- 
- {/* <CardMedia
-                className={classes.media}
-                component="img"
-                image={
-                    trackAlbums?.images[0]?.url
-                    ? trackAlbums?.images[0]?.url
-                    : DefaultImage
-                }
-                title="Album"
-              />  */}
-
-<img
-            className="Album"
-            src={trackAlbums?.images[0]?.url}
-            onError={(e) => {
-              e.target.onerror = null;
-              e.target.src = DefaultImage;
-            }}
-            alt="Album"
-          />
-
-
+        <h1>{"New Released Albums"}</h1>
         <br />
         <Grid container className={classes.grid} spacing={5}>
-          {trackAlbums?.tracks?.items.map((track) => buildCard(track))}
+          {musicAlbums?.map((album) => buildCard(album))}
         </Grid>
+        {/* <ul>
+          {musicAlbums?.map((album) => (
+            <li key={album.id}>{album.name}</li>
+          ))} 
+        </ul>*/}
       </div>
     );
 };
-export default AlbumSong;
+
+export default NewReleases;
