+++ client/src/styles/UserAuth.css
@@ -34,7 +34,6 @@
 }
 
 .card .text-field {
-  border: 1px solid rgb(188, 188, 188);
   width: 100%;
   border-radius: 4px;
   color: white;
@@ -49,7 +48,12 @@
 }
 
 .card .smallBtn {
+  padding: 0.25rem;
   color: rgb(3, 173, 215);
+}
+
+.displayBlock {
+  display: block;
 }
 
 .btn {
@@ -59,7 +63,7 @@
   font-size: 1rem;
   color: whitesmoke;
   background-color: #fa709a;
-  padding: 0.8rem 1.8rem;
+  padding: 0.25rem;
   border-radius: 2rem;
   cursor: pointer;
   margin: 24px auto;
